Trabalho 1/requirements.txt
asarPy
pycryptodomex
